For part1.py, I added three more variables, third_value, fourth_value, and fifth_value, to make them the additional three numbers the user enters. I made the variables integers like the first two variables, making the variables = int(input(“Enter a third number: “)), int(input(“Enter a fourth number: “)), and int(input(“Enter a fifth number: “)) for the third_value, fourth_value, and fifth_value variables, respectively. I then added the three variables to the first_value and second_value so that the program printed the sum of all variables.
For part2.py, I added int() to input(“Enter a number: “) for the variable number since a string, which the variable number would have been without the int(), could not be concatenated with an integer using the + operator and the program needed to find the sum of a number and 99.

For part3.py, I added int() to input(“Enter a number: “) for the variable number1 since it had to be an integer value to be used with the + operator with number2, which was an integer value. I also added a closed parenthesis ) to the end of int(input(“Enter another number: “) for number2 since number2 needed to have int() rather than int( to be an integer value. Moreover, I deleted the space between print and “The sum of your numbers is” on line 13 and added parentheses around “The sum of your numbers is”, number1 + number2 so that the program printed it. I also made Number2 number2 since Number2 was not a variable in the program, but number2 was. I also added quotes around Seven times your second number is since it was a string, and I made 7(number2) be 7 * number2 instead since * instead of parentheses was for multiplication.

For part4.py, I created two variables, base and height, and made them integers with int() and print their respective statements that told the user to input the base and height. Then I created a print() statement with the string “The area of the triangle is”, followed by (base *height)/2 that told the program to multiply the values of the variables and divide the product by two with a /, which always gave a float value.

For part5.py, I created the variables first_name and last_name that told the user to enter their first and last name. I then made a print() statement with the string “Hello,” that was followed by first_name + “ “ + last_name that told the program to concatenate the strings, including the space between the first and last names.